ui, api: move user settings view to api, cleanup
