+++ src/lando/api/views.py
@@ -9,7 +9,14 @@ from typing import Any, Callable
 from django import forms
 from django.conf import settings
 from django.core.handlers.wsgi import WSGIRequest
-from django.http import Http404, HttpRequest, HttpResponse, JsonResponse
+from django.db import IntegrityError
+from django.http import (
+    Http404,
+    HttpRequest,
+    HttpResponse,
+    HttpResponseNotAllowed,
+    JsonResponse,
+)
 from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils.decorators import method_decorator
@@ -29,6 +36,7 @@ from lando.main.models import (
     JobAction,
     JobStatus,
     LandingJob,
+    Profile,
     Repo,
     Revision,
     add_revisions_to_job,
@@ -36,6 +44,7 @@ from lando.main.models import (
 from lando.main.models.landing_job import get_jobs_for_pull
 from lando.main.models.revision import DiffWarning, DiffWarningStatus
 from lando.main.scm import SCMType
+from lando.ui.forms import UserSettingsForm
 from lando.utils.exceptions import NotFoundProblemException
 from lando.utils.github import (
     PR_DELIMITER,
@@ -574,3 +583,69 @@ class LandingJobApiView(View):
                     ]
                 }
                 return JsonResponse(data, status=400)
+
+
+@require_authenticated_user
+def manage_api_key(request: WSGIRequest) -> JsonResponse:
+    """Sets `phabricator-api-token` cookie from the UserSettingsForm.
+
+    Sets the cookie to the value provided in `phabricator_api_key` field.
+    If `reset_key` is `True` cookie is set to an empty string.
+    """
+    if not request.method == "POST":
+        return HttpResponseNotAllowed()
+
+    form = UserSettingsForm(request.POST)
+
+    if not form.is_valid():
+        return JsonResponse({"errors": form.errors}, status=400)
+
+    profile = request.user.profile
+    if form.cleaned_data["reset_key"]:
+        profile.clear_phabricator_elements()
+    else:
+        api_key = form.cleaned_data["phabricator_api_key"]
+
+        logger.debug("Verifying Phabricator API key via `user.whoami`.")
+        phab = get_phabricator_client(api_key=api_key)
+        whoami = phab.verify_api_token()
+        if not whoami:
+            return JsonResponse(
+                {"errors": {"phabricator_api_key": ["Invalid API key."]}},
+                status=400,
+            )
+
+        phid = whoami["phid"]
+        logger.debug("Phabricator API key verified for PHID `%s`.", phid)
+
+        if (
+            Profile.objects.filter(phabricator_phid=phid)
+            .exclude(pk=profile.pk)
+            .exists()
+        ):
+            return phid_conflict_response(phid)
+
+        try:
+            profile.save_phabricator_elements(api_key, phid=phid)
+        except IntegrityError:
+            return phid_conflict_response(phid)
+
+    return JsonResponse({"success": True}, status=200)
+
+
+def phid_conflict_response(phid: str) -> JsonResponse:
+    """Return a 400 response indicating the PHID is owned by another account."""
+    logger.info(
+        "Phabricator PHID `%s` is already linked to another Lando account.",
+        phid,
+    )
+    return JsonResponse(
+        {
+            "errors": {
+                "phabricator_api_key": [
+                    "This Phabricator account is already linked to another Lando user."
+                ]
+            }
+        },
+        status=400,
+    )
